Refactor reducers tests
