+++ src/App.jsx
@@ -1,15 +1,17 @@
 // @flow
 
+/* eslint-disable */
+
 /* eslint-disable import/no-named-as-default */
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import LoginView from './Components/LoginView';
 import MessageList from './Components/MessageList';
 import Toolbar from './Components/Toolbar';
+import type { State } from './FlowTypes/Types';
 import './App.css';
 
 class App extends Component {
-
   props: {
     isConnectedWithSlack: boolean,
     selectedChannel: mixed,
@@ -35,7 +37,7 @@ class App extends Component {
   }
 }
 
-export const mapStateToProps = (state) => {
+export const mapStateToProps = (state: State) => {
   const currentScore = state.scoreData[state.selectedChannel] || 0.01;
   const messages = state.channelData[state.selectedChannel] || {};
   return {
